dl_scratch4/pyproject.toml
[tool.poetry]
name = "dl_scratch4"
version = "0.1.0"
description = ""
authors = ["Joichiro433 <joichiro322@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
numpy = "^1.22.4"
pandas = "^1.4.2"
matplotlib = "^3.5.2"
scipy = "^1.8.1"
jupyterlab = "^3.4.2"
rich = "^12.4.4"
seaborn = "^0.11.2"
plotly = "^5.8.0"
coloredlogs = "^15.0.1"
scikit-learn = "^1.1.1"
opencv-python = "^4.5.5"
nptyping = "^2.1.1"
pytest = "^7.1.2"
tqdm = "^4.64.0"
# tensorflow = "^2.9.1"
tensorflow-macos = "^2.9.2"
dezero = "^0.0.13"
classopt = "^0.1.9"
pretty-errors = "^1.2.25"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
